feat: add support for email authentication without timestamp
