+++ packages/circuits/package.json
@@ -4,6 +4,7 @@
   "version": "1.0.3",
   "scripts": {
     "build": "mkdir -p build && circom src/email_auth.circom --r1cs --wasm --sym --c -l ../../node_modules -o ./build",
+    "build-no-timestamp": "mkdir -p build/no-timestamp && circom src/email_auth_no_timestamp.circom --r1cs --wasm --sym --c -l ../../node_modules -o ./build/no-timestamp",
     "build-legacy": "mkdir -p build && circom src/email_auth_legacy.circom --r1cs --wasm --sym --c -l ../../node_modules -o ./build",
     "build-recipient": "mkdir -p build && circom src/email_auth_with_recipient.circom --r1cs --wasm --sym --c -l ../../node_modules -o ./build",
     "dev-setup": "NODE_OPTIONS=--max_old_space_size=16384 npx ts-node scripts/dev-setup.ts --output ./build",
