+++ packages/circuits/src/email_auth.circom
@@ -2,4 +2,4 @@ pragma circom 2.1.6;
 
 include "./email_auth_template.circom";
 
-component main = EmailAuth(121, 17, 1024, 1024, 605, 0, 1);
+component main = EmailAuth(121, 17, 1024, 1024, 605, 0, 1, 1);
